add PA initial Dss Files Interval Dialog
