+++ wrims_v2/wrimsv2_plugin/src/wrimsv2_plugin/debugger/launcher/WPPPATab.java
@@ -34,9 +34,14 @@ import org.eclipse.swt.widgets.Display;
 import org.eclipse.swt.widgets.Group;
 import org.eclipse.swt.widgets.Label;
 import org.eclipse.swt.widgets.ProgressBar;
+import org.eclipse.swt.widgets.Shell;
 import org.eclipse.swt.widgets.Text;
+import org.eclipse.ui.IWorkbench;
+import org.eclipse.ui.PlatformUI;
 
 import wrimsv2_plugin.debugger.core.DebugCorePlugin;
+import wrimsv2_plugin.debugger.dialog.WPPCycleWreslDialog;
+import wrimsv2_plugin.debugger.dialog.WPPPAInitialDSSIntervalDialog;
 import wrimsv2_plugin.debugger.exception.WPPException;
 import wrimsv2_plugin.debugger.pa.PAProcRun;
 import wrimsv2_plugin.tools.DssOperations;
@@ -61,7 +66,6 @@ public class WPPPATab extends AbstractLaunchConfigurationTab {
 	private String startTime="";
 	private String endTime="";
 	private ProgressBar seriesPAInitPB;
-	private boolean isCreateSeriesPAInit=true;
 	private Button cancelSeriesPAInitsBut;
 	
 	public WPPPATab(WPPMainTab mainTab){
@@ -289,7 +293,16 @@ public class WPPPATab extends AbstractLaunchConfigurationTab {
 
 			@Override
 			public void widgetSelected(SelectionEvent e) {
-				createSeriesPAInits();
+				final IWorkbench workbench=PlatformUI.getWorkbench();
+				workbench.getDisplay().asyncExec(new Runnable(){
+					
+					@Override
+					public void run() {
+						Shell shell=workbench.getActiveWorkbenchWindow().getShell();
+						WPPPAInitialDSSIntervalDialog dialog= new WPPPAInitialDSSIntervalDialog(shell, mainTab, seriesPAInitPB, configuration);
+						dialog.openDialog();
+					}
+				});
 			}
 
 			@Override
@@ -317,7 +330,7 @@ public class WPPPATab extends AbstractLaunchConfigurationTab {
 
 			@Override
 			public void widgetSelected(SelectionEvent e) {
-				isCreateSeriesPAInit=false;
+				DebugCorePlugin.isCreateSeriesPAInit=false;
 			}
 
 			@Override
@@ -511,152 +524,4 @@ public class WPPPATab extends AbstractLaunchConfigurationTab {
 			dvDayCombo.add(String.valueOf(i));
 		}
 	}
-	
-	public void createSeriesPAInits(){
-		final int interval = Integer.parseInt(paStartIntervalText.getText());
-		final int startYear=Integer.parseInt(mainTab.startYearCombo.getText());
-		final int startMonth=TimeOperation.monthValue(mainTab.startMonthCombo.getText());
-		final int startDay=Integer.parseInt(mainTab.startDayCombo.getText());
-		final int endYear=Integer.parseInt(mainTab.endYearCombo.getText());
-		final int endMonth=TimeOperation.monthValue(mainTab.endMonthCombo.getText());
-		final int endDay=Integer.parseInt(mainTab.endDayCombo.getText());
-		final String initFile = mainTab.fInitFileText.getText();
-
-		Thread thread=new Thread(){
-			private int paStartYear;
-			private int paStartMonth;
-			private int paStartDay;
-
-			public void run(){
-				paStartYear=startYear;
-				paStartMonth=startMonth;
-				paStartDay=startDay;
-				
-				Calendar cal1 = new GregorianCalendar();
-		        Calendar cal2 = new GregorianCalendar();
-				cal1.set(paStartYear, paStartMonth, paStartDay);
-				cal2.set(endYear, endMonth, endDay);
-				int i = 0;
-				boolean isFirstOne=true;
-				isCreateSeriesPAInit=true;
-				
-				Display.getDefault().asyncExec(new Runnable(){
-					
-					public void run(){
-						seriesPAInitPB.setMinimum(startYear);
-						seriesPAInitPB.setMaximum(endYear);
-						seriesPAInitPB.setVisible(true);
-					}
-				});
-
-				while ((cal1.before(cal2) || cal1.equals(cal2)) && isCreateSeriesPAInit) {
-					Display.getDefault().asyncExec(new Runnable(){
-						
-						public void run(){
-							seriesPAInitPB.setSelection(paStartYear);
-							seriesPAInitPB.setVisible(true);
-						}
-					});
-					createPAInit(paStartYear, paStartMonth, paStartDay, startYear, startMonth, startDay, i*interval, isFirstOne, initFile);
-					cal1.add(Calendar.MONTH, interval);
-					paStartYear = cal1.get(Calendar.YEAR);
-					paStartMonth = cal1.get(Calendar.MONTH);
-					paStartDay = cal1.get(Calendar.DAY_OF_MONTH);
-					i++;
-					isFirstOne=false;
-				}
-				
-				if (isCreateSeriesPAInit){
-					Display.getDefault().asyncExec(new Runnable(){
-					
-						public void run(){
-							seriesPAInitPB.setSelection(endYear);
-							seriesPAInitPB.setVisible(true);
-						}
-					});
-				}
-			}
-		};
-		thread.start();
-	}
-	
-	public void createPAInit(int paStartYear, int paStartMonth, int paStartDay, int startYear, int startMonth, int startDay, int intervalMonths, boolean isFirstOne, String initFile){
-		File initDSSFile=new File(initFile);
-		if (initFile !=null){
-			if (!initDSSFile.isAbsolute()){
-				initFile=FileProcess.procRelativePath(initFile, configuration);
-				initDSSFile=new File(initFile);
-			}
-			String paInitFileName = createPAInitFileName(initFile, paStartYear, paStartMonth, paStartDay);
-			if (initDSSFile !=null){
-				File newInitDSSFile = new File(paInitFileName);
-				try {
-					FileUtils.copyFile(initDSSFile, newInitDSSFile);
-					createInitData(paInitFileName, paStartYear, paStartMonth, paStartDay, startYear, startMonth, startDay, intervalMonths, isFirstOne);
-				} catch (IOException e) {
-					WPPException.handleException(e);
-				}
-			}
-		}
-	}
-	
-	public String createPAInitFileName(String initFile, int paStartYear, int paStartMonth, int paStartDay){
-		File file = new File(initFile);
-		String fn = file.getName();
-		String fp = file.getParent();
-		String prefix="\\pa_"+paStartYear+TimeOperation.getModMonthText(paStartMonth)+TimeOperation.getModDayhText(paStartDay)+"_";
-		return fp+prefix+fn;
-	}
-	
-	public void createInitData(String paInitFileName, int paStartYear, int paStartMonth, int paStartDay, int startYear, int startMonth, int startDay, int intervalMonths, boolean isFirstOne){
-		HecDss paInitDss;
-		try {
-			paInitDss=HecDss.open(paInitFileName);
-		} catch (Exception e) {
-			WPPException.handleException(e);
-			return;
-		}
-		Vector<String> paPathList = paInitDss.getPathnameList();
-		Collections.sort(paPathList, Collections.reverseOrder());
-		int size = paPathList.size();
-		if (isFirstOne){
-			/*
-			int sYear=startYear;
-			for (int i=0; i<size; i++){
-				String path = paPathList.get(i);
-				try {
-					TimeSeriesContainer firstDc = (TimeSeriesContainer)paInitDss.get(path);
-					long st=firstDc.startTime;
-					long stms=st*60000;
-					long l1970=1440l * 25568l * 60000l;
-					long ststd=stms-l1970;
-					int sYear1=new Date(ststd).getYear()+1900;
-					if (sYear1<sYear){
-						sYear=sYear1;
-					}
-				} catch (Exception e1) {
-					WPPException.handleException(e1);
-				}
-			}
-			startTime=TimeOperation.createStartTime(sYear, 1, 1, "1DAY");
-			endTime=TimeOperation.createEndTime1(startYear, startMonth, startDay, "1DAY");
-			*/
-		}else{
-			//paInitDss.setTimeWindow(startTime, endTime);
-			for (int i=0; i<size; i++){
-				String path = paPathList.get(i);
-				try {
-					TimeSeriesContainer dc = (TimeSeriesContainer)paInitDss.get(path);
-					TimeSeriesMath tm = new TimeSeriesMath(dc);
-					//TimeSeriesMath tm1 = (TimeSeriesMath) tm.extractTimeSeriesDataForTimeSpecification("YEAR", dsy+"-"+paStartYear, true, 0, false);
-					HecMath newTm = tm.shiftInTime(intervalMonths+"MON");
-					paInitDss.write(newTm);
-				} catch (Exception e) {
-					WPPException.handleException(e);
-				}
-			}
-		}
-		paInitDss.close();
-	}
-
 }
